requirements.txt
scanpy>=1.10
anndata
numpy
pandas
scikit-learn
umap-learn
matplotlib
seaborn
scipy
h5py
transformers
tokenizers
datasets
einops
tqdm
python-pptx
ipykernel

